requirements.txt
# MCP & Web Framework
# IMPORTANT: pin mcp version — protocol has breaking changes between minor versions
mcp==1.3.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database (async)
asyncpg==0.30.0

# Vector DB
qdrant-client==1.13.2

# Cache
redis==5.2.1

# Settings & Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Embeddings
google-generativeai==0.8.3
httpx==0.28.1

# Logging
structlog==25.1.0
